refs #22 サーバ内で認証できるように
